tests/wpt/web-platform-tests/encoding-detection/uk-windows-1251.tentative.html

<!doctype html>
<title>uk windows-1251</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>Це тест на кодування символів.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1251", 'Expected windows-1251');
    }, "Check detection result");
      done();
};
</script>
